feature/04-clean-architecture Chore: Move the character adapter
Moved to the pagination folder
