+++ app/src/main/java/com/architectcoders/arquitectomarvel/ui/main/pagination/CharacterAdapter.kt
@@ -1,4 +1,4 @@
-package com.architectcoders.arquitectomarvel.ui.main
+package com.architectcoders.arquitectomarvel.ui.main.pagination
 
 import android.view.LayoutInflater
 import android.view.View
